THIRD COMMIT - WORKING MVP
This commit is the working MVP! There rae still some bugs and errors that get thrown around I'll have to look into, but most of the game-crashing stuff should be dealt with.

This commits adds dragging movement for the bucket, location checks for the bucket, and generally just a ton of bucket stuff.

Thus, MVP!

Next step is to get score, high score, and lives sorted. Then I will move on to a main menu screen and a game over screen.

Then, I wil add audio!