FOURTH COMMIT - LIFE COUNTER
This commit adds the life counter and its transparent panel, but that is about it. It cleans up some things and ensures that the model does not reference the controller, only the View, which isn't too bad--although there are exceptions
that one day I can make cleaner.

It also adds future infrastructure, for example audio loading and file reading/writing for saving high score.

Next commit will fix a small bug with the bucket controls when using right click and will add life counter dropping.