SIXTH COMMIT - MAIN MENU
This commit adds the main menu alongside the instructions text that tells the player the gist of the game and how to play. And, best of all, IT ALL WORKS!!! The instruction panel shows up when it needs to and disappears when it needs
to as well.

There is one small bug I want to fix; getting rid of setVisible in the constructors. I want to get rid of that so it doesn't do the weird flicker thing when the game starts up. Then, I control visibility through the controller class,
which I also think makes more sense and makes things a little cleaner rather than keeping track of default states.

But, next step before that will be to add the game over screen and ability to restart! Then I'll add file saving for the high score, then I will fix that bug!