EIGHTH COMMIT - Internal reorg for jar
This commit just re-organizes the assets for better use in the jar, as well as creating a manifest. This is the build and structure that actually got put into the jar (the asset loaders are also different). Otherwise, everything
is the same.

I need to really learn to package and make jar files. Intellij can do it, but I like using the console.