REFACTORED - MENU
Refactors the menu with better code organization and single-responsibility principles.
Re-organizes the Constants to be more understandable.
Adds a start menu screen, instructions screen, support for future settings screen, and start game button.

TODO: Resize text in the instructionsPanel screen and resize components in the main menu when the screen changes size.