OBJECTS - MOVEMENT OF OBJECTS
Adds the movement and proper bounds checking of the bucket object with genuinely smart and clean code that can be re-used for other projects. Encapsulated.
Adds the game loop to run the central functionality of the program; movement, redrawing, running the game. Uses a set FPS set in the constants to run each frame at 60 fps. No delta.