OBJECTS - ABSTRACT CLASS (#3)
* OBJECTS - ABSTRACT CLASS

Adds an abstract class for objects that contains their sprite and absolute positioning in the game world.

Adds a game world object that contains a measure of width and height that is used to find the scale ratio between the game world and the window.

FIX?
- May want objects to have fixed absolute width and height in the game world that is also scaled based on the screen--probably best to keep it consistent

* OBJECTS - BUCKET, DROP, GAME WORLD

Adds the bucket class, droplet class, and game world classes to represent objects in the game.

* OBJECTS - DRAWING RESIZE-SAFE BUCKET

Adds the resizing methods to the controller classes to draw a resize-safe bucket sprite to the screen that rescales its width and height and position based on the size of the screen.

Reorganizes the asset structure for better asset management and to prevent errors from file locations changing.

* OBJECTS - MOVEMENT OF OBJECTS

Adds the movement and proper bounds checking of the bucket object with genuinely smart and clean code that can be re-used for other projects. Encapsulated.
Adds the game loop to run the central functionality of the program; movement, redrawing, running the game. Uses a set FPS set in the constants to run each frame at 60 fps. No delta.

* OBJECTS - BACKGROUND + DRAWING DROPLET

Adds the drawing and setting bounds of the droplet object, randomizing its position when it is created.

Adds a background to the menu screen, same background as the game.

* OBJECTS - FINAL COMMIT

Just some minor changes. Final commit for new branch of gameplay.