GAMEPLAY - DROP TIMER
Adds a drop timer between dropping droplets so that drops start dropping after a second (60 frames) of delay. This is decreased by one frame every time a drop is created.

Adds some asset management and disposal for memory purposes, and begins small refactoring of the GUI to remove references to old screens and panels.