GAMEPLAY - COLLISION (#4)
* GAMEPLAY - COLLISION

Adds collision between the top fifth of the bucket and the droplet in the update method of the game.

* GAMEPLAY - DROP TIMER

Adds a drop timer between dropping droplets so that drops start dropping after a second (60 frames) of delay. This is decreased by one frame every time a drop is created.

Adds some asset management and disposal for memory purposes, and begins small refactoring of the GUI to remove references to old screens and panels.

* GAMEPLAY - BALANCE

Adds better balance for the speedup and delay decrement in the game.

* GAMEPLAY - SCORE LABEL

Adds a score label that tracks how many drops have been caught. Not resize-safe.

* GAMEPLAY - High Score Label

Adds high score label that tracks the score label if it exceeds the saved high score value.

* GAMEPLAY - Lives added

Adds lives and a life counter that resizes and decrements when a drop gets past the bucket. Adds game over boolean and condition for the update loop that pauses it when lives run out.

* GAMEPLAY - Game Over and Restart

Adds the game over panel (which originally I wanted to translucently stay over the game screen, but my gui code was too spaghetti to allow it easily, so later) that allows for restarting of the game with persisting high score.

* GAMEPLAY - Bugfix

Fixed the drop delay not resetting properly when restart was pressed.

* GAMEPLAY - Return to Main Menu

Final Commit.

Adds the main menu button in the game over screen for proper functionality returning to the main menu.

Time will tell if it causes a memory leak with screen assets...

* GAMEPLAY - Assets + BugFix + Memory Check

Adds the playing of sound effects including the looping music.
Fixes a bug where resetting would set the high score to the new score even if it was not greater than the previous value.

-- Checked memory and heap usage and found generally healthy behavior. Possible small creep up in memory over time but very miniscule.