REFACTORED - MENU (#1)
* REFACTORED - MENU

Refactors the menu with better code organization and single-responsibility principles.
Re-organizes the Constants to be more understandable.
Adds a start menu screen, instructions screen, support for future settings screen, and start game button.

TODO: Resize text in the instructionsPanel screen and resize components in the main menu when the screen changes size.

* MENU - CREDITS (#2)

* MENU - CREDITS

Adds a credits screen using a JTextPane which allows more advanced formatting including centered text.

Adds the actual start game method to swap to the game screen.

* MENU - SETTINGS

Adds an empty settings panel where settings will be in the future.

TODO:
1. Can abstract out some of the panels (input panel, for example, that builds a listener and builds a back button)
2. Need to figure out how to get rid of the white bar around the buttons on these panels.

* OBJECTS - ABSTRACT CLASS (#3)

* OBJECTS - ABSTRACT CLASS

Adds an abstract class for objects that contains their sprite and absolute positioning in the game world.

Adds a game world object that contains a measure of width and height that is used to find the scale ratio between the game world and the window.

FIX?
- May want objects to have fixed absolute width and height in the game world that is also scaled based on the screen--probably best to keep it consistent

* OBJECTS - BUCKET, DROP, GAME WORLD

Adds the bucket class, droplet class, and game world classes to represent objects in the game.

* OBJECTS - DRAWING RESIZE-SAFE BUCKET

Adds the resizing methods to the controller classes to draw a resize-safe bucket sprite to the screen that rescales its width and height and position based on the size of the screen.

Reorganizes the asset structure for better asset management and to prevent errors from file locations changing.

* OBJECTS - MOVEMENT OF OBJECTS

Adds the movement and proper bounds checking of the bucket object with genuinely smart and clean code that can be re-used for other projects. Encapsulated.
Adds the game loop to run the central functionality of the program; movement, redrawing, running the game. Uses a set FPS set in the constants to run each frame at 60 fps. No delta.

* OBJECTS - BACKGROUND + DRAWING DROPLET

Adds the drawing and setting bounds of the droplet object, randomizing its position when it is created.

Adds a background to the menu screen, same background as the game.

* OBJECTS - FINAL COMMIT

Just some minor changes. Final commit for new branch of gameplay.

* GAMEPLAY - COLLISION (#4)

* GAMEPLAY - COLLISION

Adds collision between the top fifth of the bucket and the droplet in the update method of the game.

* GAMEPLAY - DROP TIMER

Adds a drop timer between dropping droplets so that drops start dropping after a second (60 frames) of delay. This is decreased by one frame every time a drop is created.

Adds some asset management and disposal for memory purposes, and begins small refactoring of the GUI to remove references to old screens and panels.

* GAMEPLAY - BALANCE

Adds better balance for the speedup and delay decrement in the game.

* GAMEPLAY - SCORE LABEL

Adds a score label that tracks how many drops have been caught. Not resize-safe.

* GAMEPLAY - High Score Label

Adds high score label that tracks the score label if it exceeds the saved high score value.

* GAMEPLAY - Lives added

Adds lives and a life counter that resizes and decrements when a drop gets past the bucket. Adds game over boolean and condition for the update loop that pauses it when lives run out.

* GAMEPLAY - Game Over and Restart

Adds the game over panel (which originally I wanted to translucently stay over the game screen, but my gui code was too spaghetti to allow it easily, so later) that allows for restarting of the game with persisting high score.

* GAMEPLAY - Bugfix

Fixed the drop delay not resetting properly when restart was pressed.

* GAMEPLAY - Return to Main Menu

Final Commit.

Adds the main menu button in the game over screen for proper functionality returning to the main menu.

Time will tell if it causes a memory leak with screen assets...

* GAMEPLAY - Assets + BugFix + Memory Check

Adds the playing of sound effects including the looping music.
Fixes a bug where resetting would set the high score to the new score even if it was not greater than the previous value.

-- Checked memory and heap usage and found generally healthy behavior. Possible small creep up in memory over time but very miniscule.